Refactor to MVC code structure
