+++ Extensions.swift
@@ -1,0 +1,29 @@
+//
+//  Extensions.swift
+//  YouTube
+//
+//  Created by David Chin on 13/10/2016.
+//  Copyright © 2016 Dakerr Consulting. All rights reserved.
+//
+
+import UIKit
+
+extension UIColor {
+    static func rgb(red: CGFloat, green: CGFloat, blue: CGFloat) -> UIColor {
+        return UIColor(red: red/255, green: green/255, blue: blue/255, alpha: 1)
+    }
+}
+
+extension UIView {
+    func addConstraints(withFormat format: String, views: UIView...) {
+        var viewsDictionary = [String: UIView]()
+        for (index, view) in views.enumerated() {
+            let key = "v\(index)"
+            viewsDictionary[key] = view
+            
+            view.translatesAutoresizingMaskIntoConstraints = false // will get warning on recovery by breaking constraint without this declaration
+        }
+        
+        addConstraints(NSLayoutConstraint.constraints(withVisualFormat: format, options: NSLayoutFormatOptions(), metrics: nil, views: viewsDictionary))
+    }
+}
